fixed table-search and able to search immediately as you type
